+++ trade_core.cpp
@@ -2,14 +2,14 @@
 
 namespace trade_core
 {
-void record_wallet_snapshot(const float wallet_val_usdt, const int timestamp, float &max_wallet_val_usdt, float &max_drawdown, WalletTrace &trace)
+void record_wallet_snapshot(const double wallet_val_usdt, const int64_t timestamp, double &max_wallet_val_usdt, double &max_drawdown, WalletTrace &trace)
 {
     if (wallet_val_usdt > max_wallet_val_usdt)
     {
         max_wallet_val_usdt = wallet_val_usdt;
     }
 
-    const float pc_change_with_max = (wallet_val_usdt - max_wallet_val_usdt) / max_wallet_val_usdt * 100.0f;
+    const double pc_change_with_max = (wallet_val_usdt - max_wallet_val_usdt) / max_wallet_val_usdt * 100.0;
     if (pc_change_with_max < max_drawdown)
     {
         max_drawdown = pc_change_with_max;
@@ -19,10 +19,10 @@ void record_wallet_snapshot(const float wallet_val_usdt, const int timestamp, fl
     trace.timestamps.push_back(timestamp);
 }
 
-ResultMetrics calculate_result_metrics(const float wallet_val_usdt, const float initial_usdt, const float max_drawdown, const TradeStats &stats)
+ResultMetrics calculate_result_metrics(const double wallet_val_usdt, const double initial_usdt, const double max_drawdown, const TradeStats &stats)
 {
     ResultMetrics metrics{};
-    metrics.gain = (wallet_val_usdt - initial_usdt) / initial_usdt * 100.0f;
+    metrics.gain = (wallet_val_usdt - initial_usdt) / initial_usdt * 100.0;
 
     // A parameter set that never entered a position: win_rate would be 0/0. Report a
     // neutral, finite result instead of letting NaN reach the banner and score file.
@@ -31,13 +31,13 @@ ResultMetrics calculate_result_metrics(const float wallet_val_usdt, const float 
         return metrics;
     }
 
-    metrics.win_rate = float(stats.nb_profit) / float(stats.nb_positions_entered) * 100.0f;
+    metrics.win_rate = double(stats.nb_profit) / double(stats.nb_positions_entered) * 100.0;
 
     // "Drawdown corrected": the gain needed to recover from max_drawdown. Zero drawdown
     // makes it 0, so guard the two divisions below rather than emitting +/-inf -- an inf
     // score would win the sweep outright over every genuinely-measured candidate.
-    metrics.ddc = (1.0f / (1.0f + max_drawdown / 100.0f) - 1.0f) * 100.0f;
-    if (!(metrics.ddc > 0.0f))
+    metrics.ddc = (1.0 / (1.0 + max_drawdown / 100.0) - 1.0) * 100.0;
+    if (!(metrics.ddc > 0.0))
     {
         return metrics;
     }
@@ -47,7 +47,7 @@ ResultMetrics calculate_result_metrics(const float wallet_val_usdt, const float 
     return metrics;
 }
 
-void populate_common_result(RUN_RESULTf &result, const ResultMetrics &metrics, const float wallet_val_usdt, const float max_drawdown, const float total_fees_paid_usdt, const TradeStats &stats, const uint max_open_trades)
+void populate_common_result(RUN_RESULTf &result, const ResultMetrics &metrics, const double wallet_val_usdt, const double max_drawdown, const double total_fees_paid_usdt, const TradeStats &stats, const uint max_open_trades)
 {
     result.WALLET_VAL_USDT = wallet_val_usdt;
     result.gain_over_DDC = metrics.gain_over_ddc;
